Update dependency requirements for torch
As PyTorch 1.10 comes as the stable version of the torch framework system, along with the configuration change between 1.9 and 1.10, the code in this repo doesn't work any more. 
For future researchers who want to adopt the framework or replicate the experiment, they need to specify the highest allowed version of PyTorch and Torchvision.